requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
cryptography
python-dotenv

